MapEventInterface.isTimedEventActive u8->int return (#11, project-wide +4108)
The shared interface declared isTimedEventActive as u8-returning, so every
caller testing its result emitted a spurious clrlwi r0,r3,24 zero-extension
that target omits (target uses plain cmpwi r3,0). Retyping the vtable slot to
int return drops the clrlwi across 14 files: 18 fns improved, 6 to 100%
(largecrate_render/update, smallbasket_render, dll_F7_init, dll_CA_update,
staffactivated_spawnMapEventDebris), matched_code +4108. One fn
(pressureSwitch hagabon_update, an already-capped #108 coloring case) saw the
int-width result web perturb its coloring net-negative; isolated it with a
block-scope u8 fn-ptr cast (#57/#35) so it keeps its prior score -> zero
regressions. dol intact.

+++ include/main/mapEventTypes.h
@@ -19,7 +19,7 @@ typedef struct MapEventInterface {
     u32 (*getObjectLoadBits)(int mapId);
     u8 pad60[0x64 - 0x60];
     void (*startTimedEvent)(int eventId, f32 duration);
-    u8 (*isTimedEventActive)(int eventId);
+    int (*isTimedEventActive)(int eventId);
     f32 (*getTimedEventProgress)(int eventId);
     u8 pad70[0x74 - 0x70];
     u8 (*getPlayerNo)(void);
